Merge remote-tracking branch 'mainline/develop' into magento-review

Conflicts:
	app/code/Magento/Catalog/Block/Adminhtml/Product/Edit/Tabs.php
	app/code/Magento/Catalog/Block/Product/AbstractProduct.php
	app/code/Magento/Catalog/Controller/Adminhtml/Product.php
	app/code/Magento/Reports/Block/Adminhtml/Grid/Column/Renderer/Customer.php
	app/code/Magento/Reports/Block/Adminhtml/Grid/Column/Renderer/Product.php
	app/code/Magento/Review/Block/Adminhtml/Grid.php
	app/code/Magento/Review/Block/Product/ReviewRenderer.php
	app/code/Magento/Review/Controller/Adminhtml/Product.php
	dev/tests/integration/testsuite/Magento/Catalog/Controller/Adminhtml/Product/ReviewTest.php
	dev/tests/integration/testsuite/Magento/Wishlist/Block/AbstractTest.php

+++ app/code/Magento/Reward/Block/Checkout/Payment/Additional.php
@@ -94,10 +94,11 @@ class Additional extends \Magento\View\Element\Template
     public function getReward()
     {
         if (!$this->getData('reward')) {
-            $reward = $this->_rewardFactory->create()
-                ->setCustomer($this->getCustomer())
-                ->setWebsiteId($this->_storeManager->getStore()->getWebsiteId())
-                ->loadByCustomer();
+            $reward = $this->_rewardFactory->create()->setCustomer(
+                $this->getCustomer()
+            )->setWebsiteId(
+                $this->_storeManager->getStore()->getWebsiteId()
+            )->loadByCustomer();
             $this->setData('reward', $reward);
         }
         return $this->getData('reward');
@@ -128,8 +129,11 @@ class Additional extends \Magento\View\Element\Template
             return false;
         }
 
-        $minPointsToUse = $helper->getGeneralConfig('min_points_balance', (int)$this->_storeManager->getWebsite()->getId());
-        return (float)$this->getCurrencyAmount() > 0 && $this->getPointsBalance() >= $minPointsToUse;
+        $minPointsToUse = $helper->getGeneralConfig(
+            'min_points_balance',
+            (int)$this->_storeManager->getWebsite()->getId()
+        );
+        return (double)$this->getCurrencyAmount() > 0 && $this->getPointsBalance() >= $minPointsToUse;
     }
 
     /**
